pyproject.toml
[project]
name = "smart-core-assistant-painel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "pwlimaverde", email = "pwlimaverde@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.2",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "apscheduler==3.11.0",
    "arrow==1.3.0",
    "asgiref==3.8.1",
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "blessed==1.21.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "dataclasses-json==0.6.7",
    "distro==1.9.0",
    "django==5.2.1",
    "django-apscheduler==0.7.0",
    "django-picklefield==3.3",
    "django-q2==1.8.0",
    "django-role-permissions==3.2.0",
    "docx2txt>=0.9",
    "faiss-cpu==1.11.0",
    "firebase-admin>=6.9.0",
    "frozenlist==1.6.0",
    "greenlet==3.2.2",
    "groq>=0.26.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "idna==3.10",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "langchain-core==0.3.62",
    "langchain-groq>=0.3.2",
    "langchain-ollama>=0.3.3",
    "langchain-openai==0.3.18",
    "langchain-text-splitters==0.3.8",
    "langchain-xai==0.2.4",
    "langsmith==0.3.42",
    "markdown>=3.8",
    "marshmallow==3.26.1",
    "multidict==6.4.4",
    "mypy-extensions==1.1.0",
    "networkx>=3.5",
    "numpy==2.2.6",
    "openai==1.82.0",
    "openpyxl>=3.1.5",
    "orjson==3.10.18",
    "packaging==24.2",
    "pandas>=2.3.0",
    "propcache==0.3.1",
    "py-return-success-or-error>=0.5.2",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pypdf==5.5.0",
    "python-dateutil==2.9.0.post0",
    "python-decouple==3.8",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "redis==3.5.3",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "sqlalchemy==2.0.41",
    "sqlparse==0.5.3",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tqdm==4.67.1",
    "types-python-dateutil==2.9.0.20250516",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.1",
    "tzlocal==5.3.1",
    "unstructured>=0.17.2",
    "urllib3==2.4.0",
    "wcwidth==0.2.13",
    "xlrd>=2.0.1",
    "yarl==1.20.0",
    "zstandard==0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "autopep8>=1.6.0",
    "blue>=0.9.1",
    "django-stubs>=5.2.0",
    "isort>=6.0.1",
    "loguru>=0.7.3",
    "pytest>=8.4.0",
    "pytest-cov>=6.1.1",
    "rich>=14.0.0",
    "ruff>=0.11.13",
    "taskipy>=1.14.1",
    "types-beautifulsoup4>=4.12.0.20250516",
    "types-requests>=2.32.0.20250602",
]
doc = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.14",
    "mkdocstrings>=0.29.1",
    "mkdocstrings-python>=1.16.12",
]

[tool.autopep8]
max-line-length = 79
aggressive = 3

[tool.isort]
profile = "black"
multi_line_output = 3
trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 79

[tool.ruff]
exclude = ["__init__.py"]

[tool.hatch.build.targets.wheel]
packages = ["src/smart_core_assistant_painel"]


[project.scripts]
# Servidor Django
dev = "smart_core_assistant_painel.main:main"
start-server = "smart_core_assistant_painel.main:main"


[tool.taskipy.tasks]
# Servidor Django
dev = "python src/smart_core_assistant_painel/main.py runserver"
start = "python src/smart_core_assistant_painel/main.py runserver"
server = "python src/smart_core_assistant_painel/main.py runserver 0.0.0.0:8000"
cluster = "python src/smart_core_assistant_painel/main.py qcluster"

# Comandos Django úteis
migrate = "cd src/smart_core_assistant_painel/APP/ui && python manage.py migrate"
makemigrations = "cd src/smart_core_assistant_painel/APP/ui && python manage.py makemigrations"
createsuperuser = "cd src/smart_core_assistant_painel/APP/ui && python manage.py createsuperuser"
collectstatic = "cd src/smart_core_assistant_painel/APP/ui && python manage.py collectstatic --noinput"
shell = "cd src/smart_core_assistant_painel/APP/ui && python manage.py shell"
startapp = "cd src/smart_core_assistant_painel/APP/ui && python manage.py startapp"

# Desenvolvimento e testes
test = "pytest"
lint = "ruff check src/"
format = "ruff format src/"
type-check = "mypy src/"

# Comandos combinados
setup = "uv sync && python src/smart_core_assistant_painel/APP/ui/manage.py migrate"
dev-setup = "uv sync --dev && python src/smart_core_assistant_painel/APP/ui/manage.py migrate && python src/smart_core_assistant_painel/APP/ui/manage.py collectstatic --noinput"

# Comandos diversos
faiss_to_json = "cd src/smart_core_assistant_painel/APP/db/banco_faiss && python faiss_to_json.py"

[tool.mypy]
ignore_missing_imports = true
